mediatek/rockchip: Use PRIu64 to print uint64_t variables
Required to build for arm64 on mediatek, and let's fix the
rockchip one, while we're at it.

BUG=b:140228960
TEST=emerge-kevin-arc64 -av arc-cros-gralloc
TEST=emerge-kukui -av arc-cros-gralloc (with ARC++ 64-bit)

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/minigbm/+/1775949
Tested-by: Nicolas Boichat <drinkcat@chromium.org>
Auto-Submit: Nicolas Boichat <drinkcat@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Commit-Queue: Nicolas Boichat <drinkcat@chromium.org>

+++ rockchip.c
@@ -7,6 +7,7 @@
 #ifdef DRV_ROCKCHIP
 
 #include <errno.h>
+#include <inttypes.h>
 #include <rockchip_drm.h>
 #include <stdio.h>
 #include <string.h>
@@ -212,8 +213,8 @@ static int rockchip_bo_create_with_modifiers(struct bo *bo, uint32_t width, uint
 	ret = drmIoctl(bo->drv->fd, DRM_IOCTL_ROCKCHIP_GEM_CREATE, &gem_create);
 
 	if (ret) {
-		drv_log("DRM_IOCTL_ROCKCHIP_GEM_CREATE failed (size=%llu)\n",
-			(unsigned long long)gem_create.size);
+		drv_log("DRM_IOCTL_ROCKCHIP_GEM_CREATE failed (size=%" PRIu64 ")\n",
+			gem_create.size);
 		return -errno;
 	}
